updated main from embedded-kz

+++ fsae-vehicle-fw/src/vehicle/apps.h
@@ -3,7 +3,7 @@
 #include "bse.h"
 
 void APPS_Init();
-void APPS_UpdateData(uint32_t rawReading1, uint32_t rawReading2);
+void APPS_UpdateData(uint16_t rawReading1, uint16_t rawReading2);
 float APPS_GetAPPSReading();
 float APPS_GetAPPSReading1();
 float APPS_GetAPPSReading2();
